Refactor ladder.ts
The PHP ladder implementation has been full of tech debt for a while.

Fortunately, TypeScript makes it _much_ easier to refactor this sort of
thing. This is the sort of refactor I never would have bothered
attempting in PHP, but TS makes it really easy to clear all the tech
debt.

The idea of caching a ladder entry in a user object may be useful
in the sim server, but it was really dumb here.